[박둘내] 인구수 건물 대기열UI 2

+++ Source/CH4_Project/Private/Building/TWPopulationBuilding.cpp
@@ -160,7 +160,7 @@ void ATWPopulationBuilding::HandlePopulationQueue()
 		return;
 	}
 
-	const int8 bIncreaseSuccess = IncreasePopulationNow();
+	const int32 bIncreaseSuccess = IncreasePopulationNow();
 	if (bIncreaseSuccess == 0)
 	{
 		ResetProductionState();
